Refactor modules, and separate concerns
- Add `core.py`
- Separate JSON job seeding to `job_seeder.py`
- Add global `CRONJOBS_JSON` variable
