+++ docs/backlog.md
@@ -1,0 +1,7 @@
+# Supertask backlog
+
+## Iteration +1
+- Config: Obtain path/URL to seed file per CLI argument `--seed=cronjobs.json`
+- Config: Obtain job store schema- and table names alongside database address
+- Config: Obtain HTTP listen address
+- Add `JobExecution` subsystem derived from `django-apscheduler`
