+++ test_job.py
@@ -1,4 +1,4 @@
-from scheduler import my_job
+from jobs import my_job
 
 
 def test_my_job(mocker, capsys):
